+++ src/endgame.rs
@@ -152,7 +152,7 @@ pub struct EndgameSolver<'a> {
     klv: Box<klv::Klv>,
     board_tiles: Vec<u8>,
     racks: [Vec<u8>; 2],
-    rack_scores: [i16; 2],
+    rack_scores: [i32; 2],
     work_buffer: WorkBuffer,
 }
 
@@ -299,7 +299,7 @@ impl<'a> EndgameSolver<'a> {
             let state = &self.work_buffer.states[state_idx as usize];
             let blanked_tile =
                 state.placed_tile.tile & !((state.placed_tile.tile as i8) >> 7) as u8;
-            rack_scores[state.placed_tile.whose as usize] -= alphabet.score(blanked_tile) as i16;
+            rack_scores[state.placed_tile.whose as usize] -= alphabet.score(blanked_tile) as i32;
             state_idx = state.parent;
         }
         (rack_scores[player_idx as usize ^ 1] - rack_scores[player_idx as usize]) as f32
